Separated email subscribe form to own component and did additional cleanup
